Зазбивка кода на модули
